Re-arrange the remaining classes
