+++ java/src/main/kotlin/io/spine/validation/java/generate/OptionGenerator.kt
@@ -24,24 +24,20 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java
+package io.spine.validation.java.generate
 
-import io.spine.protodata.java.ClassName
-import io.spine.protodata.java.CodeBlock
-import io.spine.protodata.java.FieldDeclaration
-import io.spine.protodata.java.MethodDeclaration
+import io.spine.protodata.ast.TypeName
 
 /**
- * Holds all generated validation code for a specific message type.
- *
- * @property message The class name of the target message.
- * @property constraints Code blocks to be added to the `validate()` method of the message.
- * @property fields Additional class-level fields required by the validation logic.
- * @property methods Additional class-level methods required by the validation logic.
+ * Generates Java code for a specific option.
  */
-internal class MessageValidationCode(
-    val message: ClassName,
-    val constraints: List<CodeBlock>,
-    val fields: List<FieldDeclaration<*>>,
-    val methods: List<MethodDeclaration>,
-)
+internal interface OptionGenerator {
+
+    /**
+     * Generates validation code for all option applications within the provided
+     * message [type].
+     *
+     * @param type The message to generate code for.
+     */
+    fun codeFor(type: TypeName): List<FieldOptionCode>
+}
