+++ java/src/main/kotlin/io/spine/validation/java/option/PatternFieldGenerator.kt
@@ -54,18 +54,18 @@ import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_VALUE
 import io.spine.validation.java.violation.ErrorPlaceholder.PARENT_TYPE
 import io.spine.validation.java.violation.ErrorPlaceholder.REGEX_MODIFIERS
 import io.spine.validation.java.violation.ErrorPlaceholder.REGEX_PATTERN
-import io.spine.validation.java.FieldOptionCode
+import io.spine.validation.java.generate.FieldOptionCode
 import io.spine.validation.java.expression.FieldPathClass
 import io.spine.validation.java.expression.ImmutableListClass
 import io.spine.validation.java.expression.PatternClass
 import io.spine.validation.java.expression.StringClass
-import io.spine.validation.java.ValidationCodeInjector.MessageScope.message
-import io.spine.validation.java.ValidationCodeInjector.ValidateScope.parentPath
-import io.spine.validation.java.ValidationCodeInjector.ValidateScope.violations
+import io.spine.validation.java.generate.ValidationCodeInjector.MessageScope.message
+import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.parentPath
+import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.violations
 import io.spine.validation.java.violation.constraintViolation
 import io.spine.validation.java.violation.fieldPath
 import io.spine.validation.java.violation.joinToString
-import io.spine.validation.java.mangled
+import io.spine.validation.java.util.mangled
 import io.spine.validation.java.violation.templateString
 import java.util.regex.Matcher
 import java.util.regex.Pattern
