+++ java/src/main/kotlin/io/spine/validation/java/rule/GenerationContext.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java
+package io.spine.validation.java.rule
 
 import com.google.protobuf.Message
 import io.spine.protodata.ast.Field
