+++ java/src/main/kotlin/io/spine/validation/java/rule/RuleGenerator.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java
+package io.spine.validation.java.rule
 
 import io.spine.protodata.ast.MessageType
 import io.spine.protodata.ast.TypeName
@@ -37,8 +37,9 @@ import io.spine.server.query.Querying
 import io.spine.server.query.select
 import io.spine.validation.CompilationMessage
 import io.spine.validation.Rule
-import io.spine.validation.java.ValidationCodeInjector.ValidateScope.violations
-import io.spine.validation.java.rule.generatorFor
+import io.spine.validation.java.generate.FieldOptionCode
+import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.violations
+import io.spine.validation.java.generate.OptionGenerator
 import com.squareup.javapoet.CodeBlock as PoetCodeBlock
 import com.squareup.javapoet.FieldSpec as PoetField
 import com.squareup.javapoet.MethodSpec as PoetMethod
