+++ java/src/main/kotlin/io/spine/validation/java/util/JavaIdentifiers.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java
+package io.spine.validation.java.util
 
 import io.spine.protodata.backend.SecureRandomString
